Add files via upload
The current scripts still have a few soft errors but run nicely with 16 mt-plates. 
Hamilton_Data_extractor is for Hamilton data only
Clustered_luminescence is for plots
Fw/Rev_primers are the primers used
The rest are the scripts for combining the data. Fix the paths and so on in the lead script to run it.